move function from iwfpapi to top-level
