+++ firestore.rules
@@ -2,7 +2,8 @@ rules_version = '2';
 service cloud.firestore {
   match /databases/{database}/documents {
     match /{document=**} {
-      allow read, write: if false;
+      allow write: if false;
+      allow read: if true;
     }
   }
 }
