+++ functions/jest.config.js
@@ -1,6 +1,6 @@
 module.exports = {
   clearMocks: true,
-  moduleFileExtensions: ['js', 'ts', 'node'],
+  moduleFileExtensions: ['js', 'ts', 'node', 'json'],
   testEnvironment: 'node',
   testMatch: ['**/*.test.ts'],
   testRunner: 'jest-circus/runner',
