+++ functions/src/handler/get_credit_cards.ts
