+++ functions/src/handler/remove_credit_card.ts
