+++ functions/src/handler/remove_promo.ts
