+++ functions/src/handler/remove_user.ts
