+++ functions/src/util/parsers/promo.ts
