+++ functions/src/util/validate_auth.ts
