+++ functions/src/util/validate_card.ts
