+++ functions/test/fixture/promos.ts
