+++ functions/test/unit_tests/validate_auth.test.ts
