+++ functions/test/utilities/clear_database.ts
