+++ functions/test/utilities/getters/promo.ts
