+++ functions/test/utilities/validators/promo_existence.ts
