--- iwfpapi/README.md
@@ -1,7 +1,0 @@
-# iwfpapi
-
-The backend APIs for the `iwfpapp` and possibly other third party services.
-
-The root directory (here) is used for [Firebase](https://firebase.google.com/) configuration.
-
-All the API implementations are in `./functions`.
