--- iwfpapi/functions/.gitignore
@@ -1,13 +1,0 @@
-## Compiled JavaScript files
-**/*.js
-**/*.js.map
-!jest.config.js
-
-# Typescript v1 declaration files
-typings/
-
-# Third-party packages
-node_modules/
-
-# Generated documentation
-docs/
