--- iwfpapi/functions/package.json
@@ -1,47 +1,0 @@
-{
-  "name": "functions",
-  "scripts": {
-    "lint": "tslint --project tsconfig.json",
-    "fix": "tslint --fix --project tsconfig.json && prettier --write ./src/**/*.ts ./test/**/*.ts ./src/*.ts ./test/*.ts",
-    "build": "tsc",
-    "docs": "typedoc --out docs src",
-    "serve": "npm run build && firebase serve --only functions",
-    "shell": "npm run build && firebase functions:shell",
-    "start": "npm run shell",
-    "test": "jest",
-    "e2e": "npm run build && firebase emulators:exec --only functions,firestore \"npm test\"",
-    "device": "npm run build && firebase emulators:exec --only functions,firestore \"cd ../../app && flutter drive --target=test_driver/emulator.dart\"",
-    "emulator": "npm run build && firebase emulators:start --only firestore,functions",
-    "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log"
-  },
-  "engines": {
-    "node": "8"
-  },
-  "main": "lib/src/index.js",
-  "dependencies": {
-    "@types/cors": "^2.8.6",
-    "@types/sinon": "^9.0.0",
-    "cors": "^2.8.5",
-    "firebase-admin": "^8.10.0",
-    "firebase-functions": "^3.6.0"
-  },
-  "devDependencies": {
-    "@types/jest": "^25.2.1",
-    "@types/node": "^13.11.1",
-    "axios": "^0.19.2",
-    "firebase": "^7.14.0",
-    "firebase-functions-test": "^0.2.0",
-    "jest": "^25.3.0",
-    "jest-circus": "^25.3.0",
-    "prettier": "2.0.4",
-    "sinon": "^9.0.2",
-    "ts-jest": "^25.3.0",
-    "ts-mockito": "^2.5.0",
-    "ts-node": "^8.8.2",
-    "tslint": "^6.1.1",
-    "typedoc": "^0.17.4",
-    "typescript": "^3.8.3"
-  },
-  "private": true
-}
