--- iwfpapi/functions/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "noImplicitReturns": true,
-    "noImplicitAny": false,
-    "noUnusedLocals": true,
-    "outDir": "lib",
-    "sourceMap": true,
-    "strict": true,
-    "types": ["jest"],
-    "target": "es2017",
-    "esModuleInterop": true
-  },
-  "compileOnSave": true,
-  "exclude": ["node_modules"]
-}
